+++ hedera-node/hedera-smart-contract-service-impl/src/test/java/com/hedera/node/app/service/contract/impl/test/TestHelpers.java
@@ -948,10 +948,9 @@ public class TestHelpers {
     }
 
     public static org.apache.tuweni.bytes.Bytes bytesForRedirect(final byte[] subSelector, final Address tokenAddress) {
-        return org.apache.tuweni.bytes.Bytes.concatenate(
-                org.apache.tuweni.bytes.Bytes.wrap(HtsCallAttempt.LEGACY_REDIRECT_FOR_TOKEN.selector()),
-                tokenAddress,
-                org.apache.tuweni.bytes.Bytes.of(subSelector));
+        return org.apache.tuweni.bytes.Bytes.wrap(HtsCallAttempt.LEGACY_REDIRECT_FOR_TOKEN
+                .encodeCall(Tuple.of(asHeadlongAddress(tokenAddress), subSelector))
+                .array());
     }
 
     // Encode given a ByteBuffer and accountId input bytes for a call to a given contract.
@@ -963,18 +962,16 @@ public class TestHelpers {
 
     public static org.apache.tuweni.bytes.Bytes bytesForRedirectAccount(
             final byte[] subSelector, final Address accountAddress) {
-        return org.apache.tuweni.bytes.Bytes.concatenate(
-                org.apache.tuweni.bytes.Bytes.wrap(HasCallAttempt.LEGACY_REDIRECT_FOR_ACCOUNT.selector()),
-                accountAddress,
-                org.apache.tuweni.bytes.Bytes.of(subSelector));
+        return org.apache.tuweni.bytes.Bytes.wrap(HasCallAttempt.LEGACY_REDIRECT_FOR_ACCOUNT
+                .encodeCall(Tuple.of(asHeadlongAddress(accountAddress), subSelector))
+                .array());
     }
 
     public static org.apache.tuweni.bytes.Bytes bytesForRedirectScheduleTxn(
             final byte[] subSelector, final Address scheduleAddress) {
-        return org.apache.tuweni.bytes.Bytes.concatenate(
-                org.apache.tuweni.bytes.Bytes.wrap(HssCallAttempt.LEGACY_REDIRECT_FOR_SCHEDULE_TXN.selector()),
-                scheduleAddress,
-                org.apache.tuweni.bytes.Bytes.of(subSelector));
+        return org.apache.tuweni.bytes.Bytes.wrap(HssCallAttempt.LEGACY_REDIRECT_FOR_SCHEDULE_TXN
+                .encodeCall(Tuple.of(asHeadlongAddress(scheduleAddress), subSelector))
+                .array());
     }
 
     public static org.apache.tuweni.bytes.Bytes asBytesResult(final ByteBuffer encoded) {
