+++ hedera-node/hedera-smart-contract-service-impl/src/test/java/com/hedera/node/app/service/contract/impl/test/handlers/EthereumTransactionHandlerTest.java
@@ -539,7 +539,7 @@ class EthereumTransactionHandlerTest {
             given(ethTxDataReturned.to()).willReturn(toAddress);
             given(ethTxDataReturned.type()).willReturn(EthTransactionType.EIP7702);
             given(ethTxDataReturned.authorizationList()).willReturn(new byte[1]);
-            given(ethTxDataReturned.extractCodeDelegations()).willThrow(new IllegalArgumentException());
+            given(ethTxDataReturned.authorizationListAsRlp()).willReturn(new Object[] {});
             PreCheckException exception =
                     assertThrows(PreCheckException.class, () -> subject.pureChecks(pureChecksContext));
             assertEquals(INVALID_ETHEREUM_TRANSACTION, exception.responseCode());
